requirements.txt
social-auth-app-django
python-jose
python-dotenv
twilio
django-ckeditor
pillow
pytils
yandex_checkout
python-dateutil